Split on the main/utility axis
